Julia_Skvortsova/lesson2/requirements.txt
openai>=1.12.0
agents>=0.1.0
pydantic>=2.0.0
pyyaml>=6.0
python-dotenv>=1.0.0
pytest>=7.0.0
pytest-asyncio>=0.21.0
pytest-mock>=3.10.0
pandas>=2.0.0